ccc-main/requirements.txt
mysql-connector-python>=8.0.0
python-dateutil>=2.8.2
pillow>=9.0.0
matplotlib>=3.0.0
tkcalendar>=1.0.0
